A Start
Not sure if someone else has been working on this or not, but I wanted to have something to show for when we came to class.  I've made some progress and got some basic features running.  It's not perfect, so if someone else has something feel free to delete this and upload yours.  I always still have a copy on my computer.