Add initial state diagram
I made this under the assumption that our fridge would continue to cool even with the door open. However, some newer fridges nowadays stay turned off if their door is open in order to save energy. The rubric isn't completely specific on which implementation we should use, however, it does state a few things that would imply we should use my current implementation.  Just let me know if it should be changed and I'll get to it. This was made on draw.io, in case anyone wants to make edits.