Update state table and diagram
Let me know if these diagrams look accurate. I know that we have four states total in our actual code, but I'm having a hard time seeing how to implement that into the diagram. Our instructor told us that the cooler will never be activated if the door is opened, it will always be in the idle state, so I'm assuming what I have for the diagram is the most minamlistic design. Let me know if I should make other changes though.